requirements.txt
anyio==4.13.0
numpy==2.4.4
openpyxl==3.1.5
packaging==26.0
pandas==2.2.2
pluggy==1.6.0
Pygments==2.20.0
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-telegram-bot==21.4
setuptools==82.0.1
typing_extensions==4.15.0
